minor scheduler and main handler addons
some initial testing with the CLibrary type has been inserted however
some major issues regarding the requirements for final types in nested
inner classes and methos has to be solved for the TellstickScheduler to
be able to work with the Tellstick API in a correct manner. Also a minor
update in the project management: updated the Telldus daemon with codes
for the current nexa devices.