new database connector interface
a new core template regarding database connector has been added to be
able to reuse the structure in other projects. schedule has been updated
with minor issue. database structure has been updated with several major
issues regarding index and several typos. sql querys has been added to
an interface but its exprimental if the strings contained in those are
needed to be changed and if so, they will be moved to a seperate class
later on.