web service updates
after a very very loooong pause in this project the web service is being
developed and undergo a series of testing. this is the first part of
connecting xbmc to the handling server to be able to handle home
automation devices and starting with turning off/dimming. project
management is updated with new requirements.